+++ src/main/java/cat/itacademy/barcelonactiva/cognoms/nom/s05/t02/S05T02CaballeroJoan/Model/Service/Interfaces/PlayerService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface PlayerService {
     PlayerDTO save (PlayerDTO playerDTO);
     PlayerDTO update (PlayerDTO playerDTO);
-    PlayerDTO findById (long id);
+    PlayerDTO findById (int id);
     List<PlayerDTO> findAll ();
     double avgWinRate();
     PlayerDTO loser();
